fixed pedantic issue
git-svn-id: svn://svn.code.sf.net/p/peano/code/trunk/src@2986 aa977e5b-e928-4c84-aa66-c90e40304b12

+++ tarch/multicore/invasivetbb/BackgroundTasks.cpp
@@ -51,7 +51,7 @@ namespace {
 
 void tarch::multicore::spawnBackgroundTask(BackgroundTask* task) {
   if (
-    _maxNumberOfRunningBackgroundThreads==MaxNumberOfRunningBackgroundThreads::ProcessBackgroundTasksImmediately
+    _maxNumberOfRunningBackgroundThreads==static_cast<int>(MaxNumberOfRunningBackgroundThreads::ProcessBackgroundTasksImmediately)
   ) {
     task->run();
     delete task;
@@ -68,7 +68,7 @@ void tarch::multicore::spawnBackgroundTask(BackgroundTask* task) {
     (
       task->isLongRunning()
       &&
-      _maxNumberOfRunningBackgroundThreads>=MaxNumberOfRunningBackgroundThreads::DontUseBackgroundTasksForNormalTasks
+      _maxNumberOfRunningBackgroundThreads>=static_cast<int>(MaxNumberOfRunningBackgroundThreads::DontUseBackgroundTasksForNormalTasks)
     )
   ) {
     logDebug( "kickOffBackgroundTask(BackgroundTask*)", "no consumer task running yet or long-running task dropped in; kick off" );
